Rename ansible_base module to ansible_core.
