+++ src/antsibull/cli/doc_commands/devel.py
@@ -13,7 +13,7 @@ import asyncio_pool
 
 from .stable import generate_docs_for_all_collections
 from ... import app_context
-from ...ansible_base import get_ansible_base
+from ...ansible_core import get_ansible_core
 from ...collections import install_together
 from ...compat import asyncio_run
 from ...dependency_files import parse_pieces_file
@@ -57,7 +57,7 @@ async def retrieve(collections: t.List[str],
     async with aiohttp.ClientSession() as aio_session:
         async with asyncio_pool.AioPool(size=lib_ctx.thread_max) as pool:
             requestors['_ansible_base'] = await pool.spawn(
-                get_ansible_base(aio_session, '@devel', tmp_dir,
+                get_ansible_core(aio_session, '@devel', tmp_dir,
                                  ansible_base_source=ansible_base_source))
 
             downloader = CollectionDownloader(aio_session, collection_dir,
