+++ src/antsibull/cli/doc_commands/stable.py
@@ -16,7 +16,7 @@ import asyncio_pool
 from pydantic import ValidationError
 
 from ... import app_context
-from ...ansible_base import get_ansible_base
+from ...ansible_core import get_ansible_core
 from ...augment_docs import augment_docs
 from ...collections import install_together
 from ...compat import asyncio_run, best_get_loop
@@ -85,7 +85,7 @@ async def retrieve(ansible_base_version: str,
     async with aiohttp.ClientSession() as aio_session:
         async with asyncio_pool.AioPool(size=lib_ctx.thread_max) as pool:
             requestors['_ansible_base'] = await pool.spawn(
-                get_ansible_base(aio_session, ansible_base_version, tmp_dir,
+                get_ansible_core(aio_session, ansible_base_version, tmp_dir,
                                  ansible_base_source=ansible_base_source))
 
             downloader = CollectionDownloader(aio_session, collection_dir,
